refactor: MVC 패턴에 따른 폴더 생성 및 코드 파일 분리
